+++ keopscore/keopscore/mapreduce/cpu/CpuAssignZero.py
@@ -38,27 +38,27 @@ class CpuAssignZero(MapReduce, Cpu_link_compile):
 #include <vector>
 
 template < typename TYPE >
-int AssignZeroCpu_{self.gencode_filename}(size_t nx, size_t ny, std::vector< size_t > shapeout, TYPE* out,  TYPE **{arg.id}) {{
+int AssignZeroCpu_{self.gencode_filename}(signed long int nx, signed long int ny, std::vector< signed long int > shapeout, TYPE* out,  TYPE **{arg.id}) {{
     
     // for some reason the nx value is not correct for very special cases (like Zero reduction..)
     // so we compute the true value from the input shapeout...
-    size_t true_nx = 1;
-    for (size_t k=0; k<shapeout.size()-1; k++) {{
+    signed long int true_nx = 1;
+    for (signed long int k=0; k<shapeout.size()-1; k++) {{
         true_nx *= shapeout[k];
     }}
     
     #pragma omp parallel for
-    for (size_t i = 0; i < true_nx; i++) {{
+    for (signed long int i = 0; i < true_nx; i++) {{
         {outi.assign(c_zero_float)}
     }}
     return 0;
 }}
 
 template < typename TYPE >
-int launch_keops_{self.gencode_filename}(size_t nx, size_t ny, int tagI, std::vector< size_t > shapeout, TYPE *out, TYPE **arg) {{
+int launch_keops_{self.gencode_filename}(signed long int nx, signed long int ny, int tagI, std::vector< signed long int > shapeout, TYPE *out, TYPE **arg) {{
 
     if (tagI==1) {{
-        size_t tmp = ny;
+        signed long int tmp = ny;
         ny = nx;
         nx = tmp;
     }}
@@ -68,21 +68,21 @@ int launch_keops_{self.gencode_filename}(size_t nx, size_t ny, int tagI, std::ve
 }}
 
 template < typename TYPE >
-int launch_keops_cpu_{self.gencode_filename}(size_t dimY,
-                                         size_t nx,
-                                         size_t ny,
+int launch_keops_cpu_{self.gencode_filename}(signed long int dimY,
+                                         signed long int nx,
+                                         signed long int ny,
                                          int tagI,
                                          int tagZero,
                                          int use_half,
-                                         size_t dimred,
+                                         signed long int dimred,
                                          int use_chunk_mode,
                                          std::vector< int > indsi, std::vector< int > indsj, std::vector< int > indsp,
-                                         size_t dimout,
-                                         std::vector< size_t > dimsx, std::vector< size_t > dimsy, std::vector< size_t > dimsp,
-                                         size_t **ranges,
-                                         std::vector< size_t > shapeout, TYPE *out,
+                                         signed long int dimout,
+                                         std::vector< signed long int > dimsx, std::vector< signed long int > dimsy, std::vector< signed long int > dimsp,
+                                         signed long int **ranges,
+                                         std::vector< signed long int > shapeout, TYPE *out,
                                          TYPE **arg,
-                                         std::vector< std::vector< size_t > > argshape) {{
+                                         std::vector< std::vector< signed long int > > argshape) {{
 
 
     return launch_keops_{self.gencode_filename}< TYPE > (nx, ny, tagI, shapeout, out, arg);
